requirements.txt
numpy
parameterized
pika
pyjwt